Refactor: refactor interfaces and move them to a separate types folder #17
